reinserting get_variable_raw
This method was refactored in version 0.1 to internal method _get_variable_instance.

VariableInfo was moved from Parser class to Variable class.

The example SFR_Example.py was made to test this behavior but was not included into the tests. It is not yet included.
