+++ pyx2cscope/variable/variable_factory.py
@@ -3,12 +3,14 @@
 import logging
 import os
 import pickle
+from typing import Dict
+
 import yaml
 from enum import Enum
 from dataclasses import asdict
 
 from mchplnet.lnet import LNet
-from pyx2cscope.parser.elf_parser import DummyParser, VariableInfo
+from pyx2cscope.parser.elf_parser import DummyParser
 from pyx2cscope.parser.generic_parser import GenericParser
 from pyx2cscope.variable.variable import (
     Variable,
@@ -20,7 +22,7 @@ from pyx2cscope.variable.variable import (
     VariableUint8,
     VariableUint16,
     VariableUint32,
-    VariableUint64,
+    VariableUint64, VariableInfo,
 )
 
 class FileType(Enum):
@@ -185,12 +187,12 @@ class VariableFactory:
         """
         try:
             variable_info = self.parser.get_var_info(name)
-            return self._get_variable_instance(variable_info)
+            return self.get_variable_raw(variable_info)
         except Exception as e:
             logging.error(f"Error while getting variable '{name}' : {str(e)}")
 
-    def _get_variable_instance(self, var_info: VariableInfo) -> Variable:
-        """Create a variable object based on the provided address, type, and name.
+    def get_variable_raw(self, var_info: VariableInfo) -> Variable:
+        """Create a variable object based on the provided address, type, and name, defined by DataClass VariableInfo.
 
         Args:
             var_info (VariableInfo): details about the variable as name, address, type, array_size, etc.
@@ -232,7 +234,7 @@ class VariableFactory:
         }
 
         try:
-            var_type = var_info.type.lower().replace("_", "")
+            var_type: str = var_info.type.lower().replace("_", "")
             return type_factory[var_type](self.l_net, var_info.address, var_info.array_size, var_info.name)
         except IndexError:
             raise ValueError(
